Extract new note creation logic into reusable hook
Share note-creation logic between NewNoteButton and mod+shift+o hotkey
via useCreateNewNote hook. Ensures both paths apply tag frontmatter
from route/query consistently.
